Fix DeleteOrderCommand and test cases
There are some test cases that are not kept up
with the changes implemented. Additionally,
the DeleteOrderCommand was not up to date.

This may cause some unexpected behaviour when
running the application or JUnit test.

Fix the command and test cases to accomodate for
the Pair change.

This will allow the application to run as
expected.