Add test case and fix OOP.
The test case was insufficient and the codes do not adhere to SLAP.

This may cause future development to be poor due to lacking essential
test cases or inaccurate performing methods.

Fix the weak SLAP and add test case.

This will aid in future development by being creating a strong
code foundation.