Add Background Color when empty list
When the list is empty, the background color will reset to white, the
default color.

This is ugly and not the intended outcome of our application.

Add a default color for background when the list is empty.

This is more pleasant to the user's eyes, and will allow our
application to look better.