Fix EditCommand from null error
The EditCommand is throwing a null pointer exeception.

This may cause the User to be unable to editOrders.

Update the Order constructor used in the EditCommand to the newer
version.

This will allow the Person object in the Order object to be set, and thus
the User can use the command without errors.