Parser cannot parse spaces in phone number
The parser is unable to remove spaces in between numbers in the phone
number field.

This may cause inconvenience to some users who prefer to enter their
numbers with a spacing.

Edit Parser to remove all spaces in the phone parse.

This will allow users to include spaces in their phone number.