Fix Price display and values
The Price threshold was above 0, meaning that orders could not be 0.
Additionally, the decimal places were not displayed to 2 decimal
places.

User's are confused and unable to give free orders to their loyal
customers, and their decimal points would not display properly.
Additionally, one user reported that not accepting 0 would crash the
application.

Fix the Price object by formatting it's toString method to format in 2
decimal places regardless of value, and allow 0 to be an accepted
value.

This will fix the users that the users experienced, and prevent the
application from crashing.